copy: Close pipe and reap subprogram on abort
Prior to this commit, we had a number of shortcomings with respect to
subprogram orchestration, as compared to upstream Postgres:

(1) We failed to clean up the pipe file descriptors to the forked sh
executing PROGRAM on transaction abort. This meant the shell (and its
child PROGRAM) could continue to execute, even though the parent process
ERRORed out (for e.g. with a check constraint violation). This caused
reported gprestore hangs with data restores having --on-error-continue.

(2) In the instance where the PROGRAM was done by the time we
encountered the transaction abort, we would leave behind a zombie shell,
in addition to the leaked pipe fds.

(3) Failure to close file descriptors and reap shell children, coupled
with a long lived backend seeing many transaction aborts meant that we
could run into system limits for fds/process table entries.

(4) Even though we captured stderr for PROGRAM, we failed to capture the
exit code of the PROGRAM/shell. It also failed to report if the shell
was terminated by an unhandled signal.

Implementation notes:

(1) Unfortunately, since we don't use the upstream
Open/ClosePipeStream(), the fds we use aren't "managed" by the server -
i.e. they are not auto-closed on transaction abort. This also means that
these fds don't respect max_safe_fds.

(2) Even though external table infrastructure uses the same
piping/shelling mechanism, it has its own way of cleaning up at
transaction abort. See execute_abort_callback().

(3) Ideally, both external table code and COPY PROGRAM code should go
through fd.c. Added a FIXME to that effect.

(4) We now use wait_result_to_str() to know the exit code and signal
info from our shell/PROGRAM.

(5) Coverage has been added for these cases (and for COPY TO PROGRAM)

Appendix:

(1) We now get rid of the ifThrow arg to close_program_pipes(), as it
was always true.

(2) We use a more canonical implementation for reaping the shell, inside
pclose_with_stderr().

Multi-platform notice:

On some platforms, such as Rocky Linux, when coreutils is installed, an
explicit intermediate shell may not be forked between the backend and
the subprogram. The tests added send signals to all descendants of the
backend, to cater to this discrepancy. The effect is the same.

Reviewed-by: Jimmy Yih <jyih@vmware.com>