Fix agg memory limit to use PlanStateOperatorMemKB (#17598)
During the Postgres 12 merge, we fixed some references to work_mem in
the hash agg code. However, we still used the work_mem as the upper
limit for hash agg memory. This caused hash aggs to not use the available memory and spill
much more often.

if (aggstate)
{
        uint64 operator_mem = PlanStateOperatorMemKB((PlanState *) aggstate);
        if (operator_mem < strict_memlimit)
                strict_memlimit = operator_mem;
}

Note: In 6X, GPDB managed memory differently and gave much more memory
to hash aggs. See the implementation of PlanStateOperatorMemKB in 6X for
details.