Fix a dispatch issue when there's an incomplete set of mirrors
In GPDB, we could have a state in gp_segment_configuration where only part of
the primary segments have mirrors, such as during gpinitsystem or gpexpand.
However, before this commit, we assumed that we always have either only primary
segments, or a full set of primary/mirror segment pairs. In the latter state,
we organize the internal segment configuration in this way:

[primary seg0][mirror seg0][primary seg1][mirror seg1][primary seg2][mirror seg2]

And we assumed that a primary is located at "contentId * 2" in the above array.
That is wrong when we do not have the full set of mirrors, like:

[primary seg0][mirror seg0][primary seg1][primary seg2]

...accessing contentId 2 in this case would cause problem.

To avoid this issue, we could iterate the array and find the primary, but that
could have performance penalty. So instead, we now simply ignore the mirror if
we find that the cluster isn't fully-mirrored. E.g. in the above example we
would have:

[primary seg0][primary seg1][primary seg2]

... which is essentially same as a mirrorless cluster.

This can work becaues we know that the partially-mirrored configuration can only
exist during gpinitsystem where we know that no dispatch to mirror can happen.

Also convert the Assert in cdbcomponent_getComponentInfo to be an error, so
that we can catch abnormality even in release build.

An isolation2 test is added instead of gpinisystem/gpexpand test since we need
fault injector to have a more targeted test.