Run test query to validate IC proxy addresses further
ceb66bab5a17c69a637cc3f20db14873a1129215 introduced a check to validate
that gp_interconnect_proxy_addresses is configured correctly during the
init phase of gpexpand. It also added a test to assert that gpexpand
failed with an IC setup error (bind failure).

However that incorrectly depended on a later stage of the expand process
to hit that failure:

[INFO]:-Populating gpexpand.status_detail with data from database postgres
[ERROR]: gpexpand failed: Failed to setup ic_proxy interconnect ...
DETAIL:  The ic_proxy process failed to bind or listen.

Fortunately, when this scenario was run with detailed progress tracking,
generate_sql_to_populate_table_size was run, which exercises the
interconnect.

However, since 97d0fe2b81d7ef43a138f32d77c7c10bc762f9d3, simple progress
tracking is the default, which means that query exercising the
interconnect is no longer executed, hence causing gpexpand to actually
pass, instead of failing, as expected, in CI.

This commit runs a test query exercising the interconnect to make
validate_icproxy_addr() more self-serving. Also it moves the call to
this function, later on in gpexpand, just prior to setting up the
gpexpand schema (if we kept the call where it is, the test query
wouldn't exercise the interconnect setup on the new segments).