Check if gp_pause_on_restore_point_replay is an empty string
In the main redo loop, we have a conditional check on the
gp_pause_on_restore_point_replay GUC to see if we need to do
continuous WAL replay pause and resume logic. Unfortunately, the
conditional statement only checks if the GUC is NULL or not. It needs
to also check to see if the GUC value is a non-empty string or else
the logic will always get triggered since the GUC is by default an
empty string (although it would mainly just result in a quick no-op
return).